Cleaned up process logging interface
Process object now acceptions an options parameter which is a map<string,string>.  It looks for keys "stdin" and "stdout" to tee off respective logs using
the filenames supplied in the value.  If the value is the empty string the logname is generated from the process name, plus the optional "id" option.